Fix linting issue
At least on my system, I had to use python2.7 for linting, as a
result of the following error:

Traceback (most recent call last):
  File "C:\Users\Mike\Envs\kolibri\lib\site-packages\virtualenv.py", line 8, in <module>
    import base64
  File "c:\Users\Mike\Anaconda3\lib\base64.py", line 9, in <module>
    import re
  File "c:\Users\Mike\Anaconda3\lib\re.py", line 336, in <module>
    import copyreg
  File "C:\Users\Mike\Envs\kolibri\lib\site-packages\copyreg\__init__.py", line 7, in <module>
    raise ImportError('This package should not be accessible on Python 3. '
ImportError: This package should not be accessible on Python 3. Either you are trying to run from the python-future src folder or your installation of python-future is corrupted.