Remove problematic test/text input feature
I think the goal of the widget should be to simplify interaction with it.
Somehow I think there's a distinction to be made between "text changed" and
"text committed" events, i.e. pressing enter should be "committed" whereas
when input stops coming in probably "changed" is more appropriate. But I
want to limit the # of possible events emitted until such a use case
is found.