Add in styling skeleton
Make better use of Backbone's tagName and className attributes
to provide semantic tags and classes to be the target of style
rules.

Also add a little statically included css file for prototyping
the style rules -- later we may wish to swap this out for a more
elaborate system, e.g. use a css preprocessor.