Avoid using computed values in other computed values
computed values that define "set" functions aren't updated
properly when the state object is modified directly. This
wouldn't matter in production since the state is only
modified through setters -- but it has implications for
testability. Since we test by modifying state and observing
the result, bypassing the set functions, then a computed value
that depends on another computed value with a setter won't
realize that the dependent computed value has been updated.

Or so it seems.