Add all groups & ungrouped functionality
Users are considered ungrouped members of a classroom if they
have a membership with the classroom but with no groups in that
classroom. Moreover we assume that a learner will have at most
one membership in a group.

First, annotate the fetched classroom data with ungrouped learners.

Then in application logic, filter learners based on which group
is selected with edge cases when "All groups" or "Ungrouped" is
selected.