Use markdown-it rather than vue-markdown, to only enable line breaks
Some of the descriptions we scrape have newlines that we'd like to preserve. For
example, in Touchable Earth, each video an "Info" and "Transcript" field. We
scrape both fields and concatenate them together with a `\n\n` separator to render
for the description in Kolibri. By rendering with Markdown, these newlines can
be preserved, while ensuring we don't inject raw source HTML into the page (the
    Markdown parser sanitizes the input and doesn't output any input HTML).

For now we're only using the Markdown parser to render line breaks, because we
don't want to introduce too many new features at once and this is all we need at
the moment, while at the same time we have the flexibility to expand the scope
of the formatting we support.