Add contentnode_slim endpoint as lighter option
Introduce contentnode_slim endpoint and corresponding contentNodeSlim
resource which returns only the basic set of fields which are necessary
to browse the content on the frontend.

Additionally, assessmentdata key could potentially be removed but further
frontend changes are required to do so.

This change replaces calls to contentnode endpoints on the "Channels"
tab only and not on the "Recommended" tab as well.