Avoid doing extra query if file is not a thumbnail
Previously, get_storage_url would get called for each of the file
instances, now we simply check first if the file in question is even a
thumbnail, and skip that query if it's not.

NOTE: It was experimented with filtering files instances directly in the
ContentNodeSlimSerializer, but for some unknown reason response in this
context was longer than before (even though the content length was
reduced). This solution gives the best results tested so far.