Slimmer contentnode endpoint on Recommended
We introduce the ContentNodeSlimResource to the Recommended page by
replacing a subset of the contentnode endpoint API calls its slimmer
counterpart.

Specifically, we do this for the following endpoint filters:
- next_steps
- popular
- resume
- recommendations_for

IMPORTANT: this change removes part of the previous functionality:
displaying user progress on the Recommended page - this will be solved
as part of this PR by adding a separate API call to fetch the user
progress.