Enable coach content labels on ContentNodeSlim
In this commit we're adding back the support for labelling the coach
content for the contentnode_slim endpoints.

TODO: The get_num_coach_contents method needs to be optimized if
possible because it's producing n additional db queries (n = number
ContentNode objects) for each endpoint call.

Also, this change fixes the bug where we compared the results of the
gettatr call with the "if not" condition which of course can't work well
if the exists() method returns False.