Avoid descendant queries for node which are not CC
Previously we've been doing recursive descendants queries for all nodes
regardless of the fact that they're not coach content.

This change adds an early 0 value return if the passed content is
not a coach content, because due to the way in which the content
annotation works, its descendants have to be non-coach content as well.