Fix Python2 compatibility in generate_schema cmd
This fix adds special case handling the Python 2 version by opening the
json schema file in binary mode with no encoding (as the data has
already been encoded)

I'm not entirely sure why this issue arises, but from what I researched
it could be coming from the fact that some string literal which is
being used to populate the schema data has been decoded using the wrong
codec, e.g. if the text editor used to save it used different codec than
what we're using here.

In any case, this fix produces the same schema on Python 2 and 3
(tested with Python 2.7 and Python 3.6).