Do not render hidden KeenUI nav icon when it shouldn't be displayed
After custom styles updates of KeenUI toolbar introduced in
886a6d77b4aa1f8ab3856ce7b2d5f90a4e5753cd
it showed up that (newly increased) icon right margin actually affects
text to its right even when icon is not supposed to be rendered.
The reason was that KeenUI needs to be provided with `removeNavIcon`
prop to really remove the icon from DOM.
As a side effect of this change, v-show has become unnecessary.