Allow CoreBase main content to be > 1k px wide
This refactors CoreBase so that we can override the default max-width of
the component. This accommodates a design requirement in Coach reporting
that needs a review to make use of more available width.

The only side-effect was on the login page, so the 100% value was passed
to CoreBase in UserIndex.