Revert mistaken change to simpler solution
I misread (ie, didn't read all of before starting) the requirements.
The solution needed for ensuring all exams are "completed" when closed
is to make all existing ExamLogs "closed" whenever closing the exam.

This ensures that we have proper data on the frontend for displaying
the proper status of a closed exam's learners and average scores.