Refactor quiz header to commonCoach
Initially the quiz header was built in Reports, but it is used in Plan
as well. Now it has been moved to common and renamed to QuizHeader.

Now the QuizHeader has a 'dropdown' slot which inserts what is given
to it into the "options" slot of a BacklinkWithOptions component.

That is where a QuizOptionsDropdownMenu should be inserted.

Also added a validator to the "optionsFor" prop of
QuizOptionsDropdownMenu to ensure that the value is 'report' or 'plan'.

I also found that copying a quiz wasn't working due to missing new
Exam properties for dates archived and activated, so that is fixed.

Finally - this was all done because the old implementation mistakenly
removed the Copy and Delete actions from Plan > Quiz Details