Refactor BackLinkWithOptions & QuizStatus
- Quiz Status now better responsive appearance, accounting for layouts

- BackLinkWithOptions is renamed to HeaderWithOptions. The component
ultimately could be made even *more* general because it's ultimately
a wrapper for a 2 item grid which exposes two slots and plugs them into
a KGrid so that things are on the same line in layouts 8/12, but stack
at layout 4 sizing.
However, for simplicity's sake for now in this branch, this will work.

- Use ElapsedTime for "N seconds ago" rather than reimplementing it in StatusElapsedTime.