Don't use KResponsiveElementMixin in all ContentCards
Remove a piece of code that pushes `KResponsiveElementMixin`
into `ContentCard` in `<script>` part above `export default`
of `ContentCardGroupCarousel`.

Due to how files are bundled, the removed piece of code is evaluated
on the application start, no matter whether `ContentCardGroupCarousel`
is mounted or no. Consequently, `KResponsiveElementMixin` is unintentionally
used in all `ContentCards` in the whole app, not only in `ContentCardGroupCarousel`
itself. This means unnecessary performance losses whenever `ContentCard` is used
on a page because `KResponsiveElementMixin`s `ResizeSensor` causes render tree
layouts/reflows.

Regarding the intended usage for `ContentCard`s that are children
of `ContentCardGroupCarousel`, `elementWidth` and `elementHeight` provided
by `KResponsiveElementMixin` are not used in `ContentCard` at all,
and I haven't noticed any problems with the carousel during testing after
removing this code. Removed logic, including the warning in a comment,
is quite old. I assume that it might be obsolete.