Add random filter to fetchContentCollection
This patch adds an extra option to the fetchContentCollection method to
be able to do advanced queries to the API using the window.kolibri
object in the custom content renderer.

This random filter could be used to get random content, using
`order_by('?')` [1].

The filter is a CharFilter to avoid the cache, to get randomly ordered
content each query should have a different `random` parameter.

The page and pageSize parameters are also passed to the
fetchContentCollection

This patch also limit the content returned by fetchContentCollection to
the current channel to not allow to get content outside of the scope and
adds another filter to exclude topics from the list of returned nodes.

[1] https://docs.djangoproject.com/en/3.2/ref/models/querysets/#order-by