frontend: Refactor Application.get_full_url
Instead of trying and catching errors for the URL scheme, convert the
URL to tuple and obtain the scheme in get_full_url(), and call the
corresponding URL parser method. Also, let the parser return the
target URL and wrap it with the initialize URL in get_full_url
directly.

https://phabricator.endlessm.com/T32613