IE fix: Remove auto width of the KLabeledIcon
`width: auto` was breaking the title truncation
in Internet Explorer 11 since with this width settings
the title was overflowing the container even when it should
already be truncated and ellipsis was never displayed.

It seems that the main reason of using `width: auto`
was displaying the progress icon right to the title
without any gap in between. For that purpose using
`flex-grow: 0` works too and it doesn't break truncation
in IE.