API: Fix random content query
The queryset with random order does a query to the database for each
access to the QuerySet object and the consolidate method does some
manipulation with the queryset, so it's generating some inconsistence.

This patch gets the random list of ids as a fixed python List and then
creates the QuerySet object with that list of ids, so every access to
the queryset in the following serialization/consolidate methods will
have the same nodes.