CustomContent: Add descendantOf parameter to API
This patch adds a new optional parameter to the getContentByFilter API
call to be able to retrieve all descendant nodes using the MPTT model
fields.

With this new parameter it's possible to filter all the nodes below a
specific topic node, useful when combining with other filters like kind
or tags.