Update Node version reference in getting_started.rst
Also changed order of commands - I needed to run the cd command earlier in the setup, right after cloning the repo to get into the repo directory. It wasn't necessary to get into the kolibri folder within the repo, so I think that the command was misplaced in the order of instructions.