Base components, styles for AppBar, ImmsersiveToolbar, SideNav, CoreBase
This component bootstraps everything needed for **Facility** pages that include an AppBar (and, by association, a SideNav & LanguageSwitcherModal)

Prop `hideNavBar` allows us to accommodate the one place this is needed, the AllFacilitiesPage (we can't nav to users until a facility is selected)

Computed `title` returns either the given `title` prop - or it will return the correct facility label depending on if we're in a multi-facility device, in which case the selected facility's name is included in the title, or not - which just show the facilityLabel.

Introduced to encapsulate the Facility's app bar needs - particularly the sub nav when needed.

This component bootstraps everything needed for pages that include an ImmersiveToolbar

This now houses the window size based logic around the height of the top bar in a new `topBarHeight` computed property. This mixin is then pulled in where that data is needed (previously this was defined in several places)

In order to have that value available, we merge it with the responsiveWindowMIxin (you can't merge mixins by using the mixin property on the exported object so this is the way I saw it suggested when using a mixin inside a mixin, using the Vue.util.mergeOptions)

These now use the updated nav-components mixin computed property to get the consistent height depending on window size.

This has been changed to maintain backward compatibility with the above changes to avoid regressions until CoreBase has been removed entirely