Move base logic into a Core page component, wrap error pages
I initially intended on moving this into a core component if it was deemed necessary in other plugins - but when i realized that the error pages now had no AppBar, I decided to move it into a core component now.

New component AppBarCorePage
Renamed AppBarPageRoot (within Facility) to FacilityAppBarPage to clarify the purpose of it. That component now uses AppBarCorePage and houses the default "title" logic that applies to the Facility plugin alone.