Fix Kolibri blocking after starting with port 0
After the server starts, BaseServerPlugin waits for it to respond on the
expected port number. For this feature to work, we need it to see the
bind address from the associated httpserver, which is different from the
original bind address when binding to port 0.

For clarity, we will not initialize the cheroot.wsgi Server with a
bind_addr. Instead, it will be controlled by ServerPlugin itself.

With this change in place, we can remove PortCache because we no longer
need to reuse port numbers.