Install SDK as root in Docker image
There's no reason the SDK needs to be installed as the unprivileged
user. None of the debian or python packages are. This also sets the
environment variables expected by the Makefile and then there's no need
for using the `.env` file or passing variables related to it when
running docker.