Remove useless docker command
This doesn't do anything useful since the changes to the environment
only last as long as the process started by `RUN` is active.
Furthermore, the process is `sh` and not `bash`, where `source` is not
actually a builtin. If you look at the output, you'll see it produces an
error that's ignored.