Build as current user in docker
The need to copy things in and out of the container so it can run as the
real kivy user is a burden. In fact, the only reason you need to do this
is because p4a and other tools store intermediate outputs in the user's
home directory instead of in the current directory.

Instead, we can fake that by running the container as the current user
and setting the HOME environment variable to a writable location in the
container. Furthermore, we can mount a volume at this path and cache all
of those outputs.

Once the container is running as the current user, usage is much simpler
as we don't need to worry about ownership of files. The source directory
can be mounted directly into the container and made the current working
directory. Then the container environment behaves almost exactly like
running the commands directly in the current directory. We don't need a
user inside the container or have to copy the source tree into the
image.

This depends on p4a passing through the HOME environment variable during
builds. See https://github.com/kivy/python-for-android/pull/2582.