Support empty node_ids in get_import_export_nodes
When node_ids and exclude_node_ids are empty lists, we can understand
this to mean that no nodes are being selected. This is distinct from
when they are None, which means the default behaviour of selecting all
nodes. This is useful to provide a sane fallback when a manifest file
has no information about a channel.