Fix scroll bar jumping
Fixes jumping scrollbar when reaching the bottom of the page
for certain page heights. The issue is caused by the document
scroll height being changed by the sidebar's switching position
from absolute to fixed in the sticky calculation, resulting in
an endless cycle of the sticky calculation being called
and the sidepanel alternating between fixed and absolute position
over and over. Setting min width prevents this by making sure that
the document scroll height won't change on the sidebar positioning
updates.