[EPUB] Remove !important mark from text alignment
In the same line of reasoning that of the previous commit, the
text-alignment CSS selector was affecting ePub files in places
where it shouldn't, and making them left or right align text that
is supposed to be centered. We can't simply remove the CSS
selector since it would regress the small screen case.

Remove the `!important` mark from the CSS selector that affects
text alignment.