Simplify logging configuration
Since d0e15d2b7c the root logger is configured with the default handlers
and level. With that in place, there's no need to disable propagation of
loggers that use the default handlers. As noted in the documentation[1],
logger propagation only affects handlers and not level or filters.
Furthermore, any loggers with propagation enabled using the default log
level can just inherit that from the root logger. These two facts allow
most of the specific logger settings to be removed.

1. https://docs.python.org/3/library/logging.html#logging.Logger.propagate