core: Add FrontEndBaseHeadHook for injecting markup into the <head>
This allows plugins to inject arbitrary markup into the `<head>` of
every page. This is similar to what `FrontEndBaseSyncHook` allows, but
it’s limited to linking to CSS and JS files.

This will be needed by the PWA plugin to add a `<link>` to the PWA
manifest for the site.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>