core: Allow specifying additional logos in the theme
Add a `logos` array for defining zero or more logos to be used on the
site. These will be used in an upcoming commit for exposing the site’s
logo in various sizes in a PWA manifest.

Add additional `content_type` and `maskable` properties to the image
spec to help with this. Both are optional. `content_type` specifies the
content type of the image (such as `image/png`). `maskable` specifies
whether the image is layed out so that it can be masked by a client (see
https://web.dev/maskable-icon/).

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>