pwa: Add a Progressive Web App (PWA) plugin
This allows installing a Kolibri instance as a web app on the user’s
desktop, if their web browser supports it (Chrome does) and if the site
is configured to allow it (HTTPS must be enabled, it must be accessible
without an internet connection, and the theme must have appropriate
icons defined).

There are two main components to making an installable PWA: a manifest
and a service worker. See https://www.w3.org/TR/appmanifest/ or
https://developer.mozilla.org/en-US/docs/Web/Progressive_web_apps for an
overview of PWAs.

We serve the manifest from `/${lang}/pwa/manifest.json`, and it’s linked
to from the `<head>` of every page so that the browser can find it. It’s
localised using the normal URI i18n functionality. The web app manifest
specification doesn’t allow for localisation within the manifest file.

The service worker has to be served from `/sw.js` (not
`/${lang}/pwa/sw.js`) because its URI path determines its scope. If it
were in a nested path, it could only act on requests at that path or
beneath, which wouldn’t work. See
https://www.w3.org/TR/service-workers/#path-restriction.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>