docs: Add a section on copyright and licensing
Including an example of a minimal file header which explicitly states
the copyright and license for that file. Both can differ across the
project:
 * Copyright can vary between files because they contain contributions
   from different individuals or organisations.
 * License can vary between files if code has been brought in from
   compatibly-licensed external projects.

Copyright holdings cannot reliably be extracted from git version history
for a file, because not all contributions are significant enough to have
a copyright claim; and the author of a commit may not be the copyright
holder of what they’re committing (it might be their employer, for
example).

If the license and copyright are explicitly stated in each file, that
avoids ambiguity about who the copyright holders for that code are, and
allows for the code to be copied to other compatibly-licensed projects
without being separated from the licensing information in `./LICENSE`.

However, this change explicitly does not make that a project-wide
requirement, since that’s quite a bit workflow change and imposition on
contributors. The intent here is to document what’s possible and what
might be found in some files.

See https://reuse.software/faq/ for some general discussion of this
topic.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>