utils: Don’t initialise hooks when iterating plugins
The PWA plugin registers a hook which is used on every page. The
`test_key_urls` test opens all the URLs in the main URL map, which
includes `/plugins/` and then `/learn/`. If iterating plugins
initialises hooks, and the PWA plugin is disabled, then loading
`/plugins/` causes the PWA’s hook to be registered, and then loading
`/learn/` causes it to be executed — but the plugin is not enabled. This
causes an exception.

This doesn’t change the behaviour of `kolibri plugin list` or the
`/plugins/` endpoint on a normal installation.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>