Allow requesting all thumbnails in import/export data
Add an `all_thumbnails` keyword argument to `get_import_export_data` to
request that all thumbnails in a channel should be included and not just
those associated with the desired nodes and their associated topics.
This can be useful for the frontend to display content that is available
to download in a richer way. The option default to `False` to preserve
the existing behavior.