default_theme: Move favicon location to be configured by the theme
Previously it was hardcoded as `static/logo.ico`, which was not
overrideable by individual Kolibri installations.

Add it to the base page template using a new `theme_favicon` tag, since
getting the theme and looking up the right icon from the logos list is
not straightforward enough to do within a normal Django template
variable substitution.

We considered implementing it as a new `FrontEndBaseHeadHook` subclass,
but using that in `unsupported_browser.html` could lead to unexpected
broken JavaScript being pulled into that page in future, without anyone
noticing. Hence the use of a more constrained tag.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>