server: Add systemd notify support for process management
When running Kolibri under systemd, there’s currently no way of telling
when the Kolibri service has fully started up and opened all of its
sockets and registered successfully on zeroconf.

This means that when running Kolibri under systemd (for example as
`kolibri services --foreground`), systemd can only assume that Kolibri
is ready to handle clients as soon as fork/exec has completed. That’s
not true; it takes a couple of seconds to do the zeroconf registration.

Add `NOTIFY_SOCKET` support as a magicbus plugin to send state messages
to systemd to tell it when Kolibri is ready after starting up, and when
it’s starting to shut down.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>