options: Clarify that zipcontent should be served from a different origin
Zipcontent is not necessarily fully trusted, but it is shown in an
iframe within the main Kolibri app. If it comes from the same origin as
the Kolibri app, it will have permissions within the browser to access
the parent frame and its resources. That’s undesirable.

Clarify that zipcontent should therefore be served from a different
origin. For a Kolibri instance which is running locally, this is handled
by serving zipcontent from a different port. For a Kolibri instance
which is running online, it should be handled by serving zipcontent from
a different domain or subdomain.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>