Bypass browser cache for forced resource fetches
If the app makes a forced resource fetch, ensure the request reaches the
backend without using the browser cache by setting the `Cache-Control`
header to `max-age=0`. Note that `max-age=0` is preferred to `no-cache`
since the browser will still add an `If-No-Match` header that allows the
backend to return a 304 with no content and avoid the data transfer.