Always update content cache key after channel import
When a channel is imported, the content cache key needs to be updated so
clients won't receive stale content from `ChannelMetadataViewSet`.
Currently that only happens if the channel has content nodes, but that
would preclude a client from discovering a new channel and offering it
as available to download resources for. Clients can still query the
`ChannelMetadataViewSet` endpoint with the `available` parameter if they
prefer to only see channels that already have resources.