Stop caching content API responses
Currently channel and content node API responses are cached by Django to
avoid expensive queries. Since the Django cache doesn't respect
`Cache-Control` request headers, the client will continue to receive
cached responses until they expire or the backend removes the stale
entries.

To workaround this, the frontend includes a `contentCacheKey` query
parameter using the current `ContentCacheKey` value exported by the
backend. This changes the URL when the content cache key is updated and
bypasses both the browser and server caches. However, the cache key
value is provided to the frontend statically in the HTML `head` element.
Unless the frontend reloads the page, it will continue to use an old
`contentCacheKey` query parameter and receive cached responses from the
server. If the frontend wants to offer content downloads and update the
page when the content has changed, it has no choice but to reload the
page to ensure it receives current data from the API.

Since the caching can't be made to work in all cases, stop using it. The
content cache key is still used as the ETag, which allows the frontend
(browser) to send an `If-None-Match` HTTP header and the backend to
respond with a 304 Not Modified. That ensures the API responses will
always be current but avoids the database queries if the content cache
key hasn't changed

Fixes: #10968