kolibri-tools: Prepend to Python path rather than append
If the kolibri-tools are supposed to be looking in a certain directory
for plugins, that directory needs to be listed first in the Python path,
rather than last. Otherwise the tools could end up finding the plugin in
a system directory instead — which will probably be the wrong version of
it, or the wrong place to write out built files.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>