kolibri-tools: Add support for pyproject.toml
kolibri-tools currently hard-codes a search for a Python-style `setup.cfg`
in the project’s directory, to load locale configuration from.

As per PEP-621, `pyproject.toml` is now a recommended replacement for
`setup.cfg`, though both are going to be supported for a while. A number
of tools now support `pyproject.toml`, and others don’t.

This commit adds support for `pyproject.toml` to `kolibri-tools`, using
it in preference to `setup.cfg`, but falling back to the latter if the
former doesn’t exist. It doesn’t attempt to load some keys from one file
and some from the other: it’s all or nothing.

This allows for i18n configuration to now be specified in
`pyproject.toml` something like this:
```
[tool.kolibri.i18n]
project = "kolibri_explore_plugin"
locale_data_folder = "kolibri_explore_plugin/locale"
\# Glob to exclude node_modules and static folders
ignore = ["**/node_modules/**", "**/static/**"]
```

This differs slightly from the equivalent in `setup.cfg`, as:
 - `pyproject.toml` conventions require a section name prefixed by
   `tool.`.
 - TOML allows for more strongly typed values.

It adds a dependency on the node `toml` module.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://github.com/learningequality/kolibri/issues/11146