Show app and daemon details in the About window
To provide a single source of information about the app, this commit
adds debug info to the About window in the form of a JSON object dump.
It includes the version of the application itself, the versions of
essential Python modules, and runtime information about the Kolibri
daemon service.

Helps: https://github.com/endlessm/endless-key-flatpak/issues/69