Don't require channel name in content tasks API
It's nice to include the channel name in the tasks extra metadata so
that it can be shown in the frontend, but often the caller only has the
channel ID and is required to supply a bogus name. Instead, default to
the empty string and let anyone inspecting the task decide how they want
to deal with a missing name.