ci: Use eos-shell-build token for f-e-d-c PR
The PRs created by `flatpak-external-data-checker` currently do not
trigger the PR checks to run. This is because the default `GITHUB_TOKEN`
used to create the PR cannot trigger a recursive GitHub Actions
workflow. That's a safety measure to prevent infinitely recursing
workflows.

Instead, use a personal access token (PAT) from the eos-shell-build user
to create the PR. This token was created specifically for use in GitHub
Actions with `repo` scope. The token has been recorded in the
`WORKFLOW_TOKEN` secret.