Fix #9067: Redirect signed-out users to resource URL without lessonId or classId
With the setting "Allow users to explore resources without signing in" enabled, if a signed-out user accesses a resource page with the lessonId or classId query parameters, it auto-redirects to the same page of the topic that does not contain these parameters.

Giving access to lesson resources for users that are not signed in can cause potential problems related to lesson/topic contexts, therefore, it's better to circumvent this problem by automattically redirecting these users to an URL that doesn't contain this context.

This commit aims to fix the problem described in issue #9067.