properly handle resource_pool, question and question_count updates
this was kind of a pain to work out, ultimately, the method to select random questions
needed to take a parameter for the resource_pool it will select from.
this implementation results in a massive updateSection method. an ideal alternative
would probably involve changes to the API overall, which would require a moderate refactor
but would also likely result in a much cleaner interface