UserTable: tests ignored, proactive regression avoidance around KCheckbox
KCheckbox no longer maintains its own internal state, rather, it is just
like a fancy button which emits an event that it's been clicked and the
parent component must react and update the value it passes to `checked`
or `indeterminate` accordingly.

These tests rely on KCheckbox to emit its internal state whenever it
emits `change` and that no longer is the case given KCheckbox doesn't
keep an internal state.

I've tweaked some of the conditionals in UserTable that relied on that
emitted `checked` value so that they instead get the same information
from existing available values.