Defer Kolibri service start to next reboot on "Run on start" toggle
Previously, when the "Run Kolibri service on start" option was toggled on in the Windows taskbar icon menu, Kolibri would configure the service for automatic startup and attempt to start it.

With this change, the "Run Kolibri service on start" option only sets the Kolibri Windows service's startup type to "Automatic". The service will then start on the next system reboot.